lots

+++ src/pwm.c
@@ -5,7 +5,7 @@ void pwm_Init(struct PWM_PIN pwmPin){
   pwm_DutyCycle(pwmPin, 0);
 }
 
-void pwm_DutyCycle(struct PWM_PIN pwmPin, size_t dutyCycle){
+void pwm_DutyCycle(struct PWM_PIN pwmPin, uint16_t dutyCycle){
   timerStart(pwmPin.timer, TIMER_MODE_PWM, pwmPin.channel, PRESCALER_NONE, dutyCycle);
 }
